feat: petit changement

+++ ApiV1ControlleurMonstre/ApiV1ControlleurMonstre/Services/QueteService.cs
@@ -2,6 +2,7 @@
 using ApiV1ControlleurMonstre.Data.Context;
 using ApiV1ControlleurMonstre.Models;
 using ApiV1ControlleurMonstre.Utilities;
+using Humanizer.Localisation.DateToOrdinalWords;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.VisualBasic;
 using System.Threading;
@@ -35,43 +36,48 @@ namespace ApiV1ControlleurMonstre.Services
                 if (allQuests.Where(quest => quest.Personnage == character).ToList().Count() < 3)
                 {
                     // Sinon, généré une quest
-
-                    switch (Random.Shared.Next(3))
+                    do
                     {
-                        // Destination
-                        case 0:
-                            // Get une tuile au hazard
-                            List<Tuile> listTuilesValide = await context.Tuiles
-                                .Where(tuile =>tuile.EstTraversable == true)
-                                .ToListAsync(cancellationToken);
+                        switch (Random.Shared.Next(3))
+                        {
+                            // Destination
+                            case 0:
+                                // Get une tuile au hazard
+                                List<Tuile> listTuilesValide = await context.Tuiles
+                                    .Where(tuile => tuile.EstTraversable == true)
+                                    .ToListAsync(cancellationToken);
 
-                            Tuile destination = listTuilesValide[Random.Shared.Next(listTuilesValide.Count())];
+                                Tuile destination = listTuilesValide[Random.Shared.Next(listTuilesValide.Count())];
 
-                            // Créee la quete
-                            _logger.LogInformation($"Creating a new tile quest for {character.Name}");
-                            Quete queteDestination = new Quete($"Rendez-vous sur la tuile {destination.PositionX}, {destination.PositionY}", destination, character);
-                            context.Quetes.Add(queteDestination);
-                            break;
-                        // Monstre
-                        case 1:
-                            // Prendre un monstre au hazard 
-                            Monstre monstre = context.Monstre.ElementAt(Random.Shared.Next(context.Monstre.Count()));
-                            // Prendre le type
-                            string type = Random.Shared.Next(2) == 1 ? monstre.Type1 : monstre.Type2 ?? monstre.Type1;
+                                // Créee la quete
+                                _logger.LogInformation($"Creating a new tile quest for {character.Name}");
+                                Quete queteDestination = new Quete($"Rendez-vous sur la tuile {destination.PositionX}, {destination.PositionY}", destination, character);
+                                context.Quetes.Add(queteDestination);
+                                allQuests.Add(queteDestination);
+                                break;
+                            // Monstre
+                            case 1:
+                                // Prendre un monstre au hazard 
+                                Monstre monstre = context.Monstre.ElementAt(Random.Shared.Next(context.Monstre.Count()));
+                                // Prendre le type
+                                string type = Random.Shared.Next(2) == 1 ? monstre.Type1 : monstre.Type2 ?? monstre.Type1;
 
-                            _logger.LogInformation($"Creating a new monster quest for {character.Name}");
-                            Quete queteMonstre = new Quete($"Tuer des monstres de type {type}", type, Random.Shared.Next(11), character);
-                            context.Quetes.Add(queteMonstre);
-                            break;
-                        case 2:
-                            _logger.LogInformation($"Creating a new level quest for {character.Name}");
-                            Quete queteNiveau = new Quete($"Augmentez au niveau {character.Niveau + 1}", character);
-                            context.Quetes.Add(queteNiveau);
-                            break;
-                    }
+                                _logger.LogInformation($"Creating a new monster quest for {character.Name}");
+                                Quete queteMonstre = new Quete($"Tuer des monstres de type {type}", type, Random.Shared.Next(11), character);
+                                context.Quetes.Add(queteMonstre);
+                                allQuests.Add(queteMonstre);
+                                break;
+                            case 2:
+                                _logger.LogInformation($"Creating a new level quest for {character.Name}");
+                                Quete queteNiveau = new Quete($"Augmentez au niveau {character.Niveau + 1}", character);
+                                context.Quetes.Add(queteNiveau);
+                                allQuests.Add(queteNiveau);
+                                break;
+                        }
+                    } while (allQuests.Where(quest => quest.Personnage == character).ToList().Count() < 3);
                 }
+            await context.SaveChangesAsync(cancellationToken);
             }
-           await context.SaveChangesAsync(cancellationToken);
         }
 
         private async Task VerifyCompletion(MonsterContext context, CancellationToken cancellationToken)
@@ -86,14 +92,17 @@ namespace ApiV1ControlleurMonstre.Services
 
             // Retirer les quetes qui sont complété
             _logger.LogInformation($"Removing all completed quests");
-            context.RemoveRange(allQuests.Select(item => item.EstCompleter));
+            if (allQuests.Any(item => item.EstCompleter == true))
+            {
+                context.RemoveRange(allQuests.Select(item => item.EstCompleter));
            
-            await context.SaveChangesAsync(cancellationToken);
+                await context.SaveChangesAsync(cancellationToken);
+            }
         }
 
         protected async  override Task ExecuteAsync(CancellationToken stoppingToken)
         {
-            TimeSpan _checkInterval = TimeSpan.FromMinutes(1); // Check every 10 minutes
+            TimeSpan _checkInterval = TimeSpan.FromMinutes(10); // Check every 10 minutes
 
             using var scope = _serviceProvider.CreateScope();
             var context = scope.ServiceProvider.GetRequiredService<MonsterContext>();
@@ -109,9 +118,9 @@ namespace ApiV1ControlleurMonstre.Services
                     context.ServiceTimestamps.Add(serviceTimestamp);
                     await context.SaveChangesAsync();
 
-                    await Task.Delay(_checkInterval, stoppingToken);
                     await VerifyCompletion(context, stoppingToken);
                     await RegenerateQuest(context, stoppingToken);
+                    await Task.Delay(_checkInterval, stoppingToken);
                 }
                 catch (OperationCanceledException)
                 {
